requirements.txt
flask
flask-cors
werkzeug
pyjwt
psycopg2-binary
gunicorn
